# specs/openapi/swagger.yml
openapi: "3.0.0"
info:
  version: 1.0.1
  title: "PrimeDividend"
  description: "API для приложения PrimeDividend"
...
servers:
  - url: http://localhost/api
    description: "Local server"

#security:
#  - BearerAuth: [ ]
...
paths:
...
  /auth/{network}:
    get:
      tags: [ auth ]
      operationId: authNetwork
      summary: Аутентификация в соц. сетях
      parameters:
        - $ref: "#/components/parameters/network"
      responses:
        301:
          $ref: '#/components/responses/301'
        500:
          $ref: '#/components/responses/500'
  /auth/{network}/callback:
    get:
      tags: [ auth ]
      operationId: authNetworkCallback
      summary: Авторизация в соц. сетях
      parameters:
        - $ref: "#/components/parameters/network"
      responses:
        301:
          $ref: '#/components/responses/301'
        500:
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    network:
      in: path
      name: network
      schema:
        type: string
      required: true
  responses:
...
    301:
      description: Moved Permanently
...
    500:
      description: Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
  schemas:
...
    error:
      type: object
      required: [ errors ]
      properties:
        data:
          nullable: true
          default: null
        errors:
          type: array
          items:
            type: object
            required: [ message ]
            properties:
              message:
                type: string
                description: Описание ошибки
              field:
                type: string
                description: Поле в котором произошла ошибка